+++ docs/changelog.rst
@@ -12,6 +12,7 @@ CHANGELOG
 **Minor improvements**
 
 - Add ratings_description field in export (#2755)
+- Add accessibility field on Infrastructure
 
 **Bug fixes**